to c

+++ resources/tools/fs/flashFs.sh
@@ -19,7 +19,15 @@ mv ./littlefs/other/yatchy-lp-program*.bin ./littlefs/other/yatchy-lp-program.bi
 rm -rf /tmp/littlefs
 cp -r littlefs /tmp/
 
-find /tmp/littlefs -type f ! -path "*/book/*" ! -name ".keep" -exec python3 ./compressFile.py {} \;
+# Compile the C compressor
+echo "Compiling compressFile.c..."
+gcc -o in/compressFile -I../../../components/tamp/ ../../../components/tamp/common.c ../../../components/tamp/compressor.c compressFile.c
+if [ $? -ne 0 ]; then
+    echo "Error compiling compressFile.c. Exiting."
+    exit 1
+fi
+
+find /tmp/littlefs -type f ! -path "*/book/*" ! -name ".keep" -exec bash -c './in/compressFile "$0" "$0.compressed" && mv "$0.compressed" "$0"' {} \;
 rm -rf out/fs.bin
 ./in/mklittlefs --all-files -c /tmp/littlefs -s $size out/fs.bin
 rm -rf /tmp/littlefs
